requirements.txt
numpy
opencv-python
pyds
openvino-dev
gobject-introspection
pipreqs
networkx
defusedxml
openvino-telemetry
packaging